Simplify derive_nullifier
